Resolve "[Plugin System] Add introspection to the dicey server"

Closes #18

See merge request zuru.tech/dreamcatcher/platform/dicey!19

+++ src/ipc/typedescr.c
@@ -36,7 +36,7 @@ static bool match_exact(const char **const cur, const char *const end, const cha
 static void skip_whitespace(const char **const cur, const char *const end) {
     assert(cur && *cur && end);
 
-    while (*cur != end && isspace(**cur)) {
+    while (*cur != end && isspace((unsigned short) **cur)) {
         ++*cur;
     }
 }
